Fix country dropdown
