+++ src/common/helpers/getLanguageCode.ts
@@ -1,0 +1,11 @@
+function getLanguageCode(langOrLangAndLocale: string) {
+  const hasLocale = langOrLangAndLocale.includes('-');
+
+  if (hasLocale) {
+    return langOrLangAndLocale.split('-')[0];
+  }
+
+  return langOrLangAndLocale;
+}
+
+export default getLanguageCode;
